+++ CarSelling/Models/IEngine.cs
@@ -3,7 +3,7 @@
     public interface IEngine
     {
         public FuelType FuelType { get; set; }
-        public ushort EngineCapacity { get; set; }
+        public double EngineCapacity { get; set; }
         public ushort Power { get; set; }
     }
 }
